+++ demos/27-polygon-intersection-greinerhormann/drawTriangulation_delaunay.js
@@ -5,40 +5,66 @@
  * * Convex polygons are fine
  * * Self intersections are fine
  *
+ * @param {PlotBoilerplate} pb
  * @param {Polygon} intersectionPolygon
  * @param {Polygon} sourcePolygon
  * @param {Polygon} clipPolygon
+ * @param {boolean} drawDelaunayCircles
+ * @return {Triangle[]}
  */
-var drawTriangulation_delaunay = function( pb, intersectionPolygon, sourcePolygon, clipPolygon, drawDelaunayCircles ) {
-    var selfIntersectionPoints = findPolygonSelfIntersections( intersectionPolygon );
-    var extendedPointList = intersectionPolygon.vertices.concat( selfIntersectionPoints );
 
-    var delaunay = new Delaunay( extendedPointList, {} );
-    // Array<Triangle>
-    var triangles = delaunay.triangulate();
+(function(context) {
+    
+    var drawTriangulation_delaunay = function( pb,
+					       intersectionPolygon,
+					       sourcePolygon,
+					       clipPolygon,
+					       drawDelaunayCircles
+					     ) {
+	var selfIntersectionPoints = findPolygonSelfIntersections( intersectionPolygon );
+	var extendedPointList = intersectionPolygon.vertices.concat( selfIntersectionPoints );
 
-    // Find real intersections with the triangulations and the polygon
-    // extendedPointList    
+	var delaunay = new Delaunay( extendedPointList, {} );
+	// Array<Triangle>
+	var triangles = delaunay.triangulate();
 
-    // Remember: delaunay returns an array with lots of empty slots.
-    //           So don't use triangles.length to access the triangles.
-    for( var i in triangles ) {
-	var tri = triangles[i];
-	// Check if triangle belongs to the polygon or is outside
-	if( !sourcePolygon.containsVert( tri.getCentroid() ) )
-	    continue;
-	if( !clipPolygon.containsVert( tri.getCentroid() ) )
-	    continue;
+	// Find real intersections with the triangulations and the polygon
+	// extendedPointList    
 
-	// Cool, triangle is part of the intersection.
-	pb.draw.polyline( [tri.a, tri.b, tri.c], false, 'rgba(0,128,255,0.5)', 1 );
-	// drawFancyCrosshair( pb, tri.a, false, false );
-	// drawFancyCrosshair( pb, tri.b, false, false );
-	// drawFancyCrosshair( pb, tri.c, false, false );
-	if( drawDelaunayCircles ) {
-	    var circumCircle = tri.getCircumcircle();
-	    pb.draw.crosshair( circumCircle.center, 5, 'rgba(255,0,0,0.25)' );
-	    pb.draw.circle(  circumCircle.center, circumCircle.radius, 'rgba(255,0,0,0.25)',1.0 );
+	// Remember: delaunay returns an array with lots of empty slots.
+	//           So don't use triangles.length to access the triangles.
+	for( var i in triangles ) {
+	    var tri = triangles[i];
+	    // Check if triangle belongs to the polygon or is outside
+	    if( !sourcePolygon.containsVert( tri.getCentroid() ) )
+		continue;
+	    if( !clipPolygon.containsVert( tri.getCentroid() ) )
+		continue;
+
+	    // Cool, triangle is part of the intersection.
+	    pb.draw.polyline( [tri.a, tri.b, tri.c], false, 'rgba(0,128,255,0.5)', 1 );
+	    // drawFancyCrosshair( pb, tri.a, false, false );
+	    // drawFancyCrosshair( pb, tri.b, false, false );
+	    // drawFancyCrosshair( pb, tri.c, false, false );
+	    if( drawDelaunayCircles ) {
+		var circumCircle = tri.getCircumcircle();
+		pb.draw.crosshair( circumCircle.center, 5, 'rgba(255,0,0,0.25)' );
+		pb.draw.circle(  circumCircle.center, circumCircle.radius, 'rgba(255,0,0,0.25)',1.0 );
+	    }
 	}
-    }
-};
+	
+	return filterNonNull(triangles);
+    };
+
+    var filterNonNull = function( arr ) {
+	var filtered = [];
+	for( var i = 0; i < arr.length; i++ ) {
+	    if( typeof arr[i] !== 'undefined' )
+		filtered.push( arr[i] );
+	}
+	return filtered;
+    };
+    
+    context.drawTriangulation_delaunay = drawTriangulation_delaunay;
+
+})(globalThis || window);