+++ src/js/Triangle.js
@@ -16,8 +16,9 @@
  * @modified  2020-05-12 Added getIncircularTriangle() function.
  * @modified  2020-05-12 Added getIncircle() function.
  * @modified  2020-05-12 Fixed the signature of getCircumcirle(). Was still a generic object.
- * @modified  2020-06-18 Added the getIncenter function.
- * @version   2.3.0
+ * @modified  2020-06-18 Added the `getIncenter` function.
+ * @modified  2020-12-28 Added the `getArea` function.
+ * @version   2.4.0
  *
  * @file Triangle
  * @fileoverview A simple triangle class: three vertices.
@@ -80,11 +81,25 @@ var Triangle = /** @class */ (function () {
      * @return {Triangle}
      **/
     Triangle.fromArray = function (arr) {
-        //if( !Array.isArray(arr) )
-        //    throw new Exception("Cannot create triangle fromArray from non-array.");
         if (arr.length < 3)
             throw "Cannot create triangle from array with less than three vertices (" + arr.length + ")";
         return new Triangle(arr[0], arr[1], arr[2]);
+    };
+    ;
+    /**
+     * Get the area of this triangle. The returned area is never negative.
+     *
+     * If you are interested in the signed area, please consider using the
+     * `Triangle.utils.signedArea` helper function. This method just returns
+     * the absolute value of the signed area.
+     *
+     * @method getArea
+     * @instance
+     * @memberof Triangle
+     * @return {number} The non-negative area of this triangle.
+     */
+    Triangle.prototype.getArea = function () {
+        return Math.abs(Triangle.utils.signedArea(this.a.x, this.a.y, this.b.x, this.b.y, this.c.x, this.c.y));
     };
     ;
     /**
@@ -399,6 +414,9 @@ var Triangle = /** @class */ (function () {
         min3: function (a, b, c) {
             return (a <= b && a <= c) ? a : (b <= a && b <= c) ? b : c;
         },
+        signedArea: function (p0x, p0y, p1x, p1y, p2x, p2y) {
+            return 0.5 * (-p1y * p2x + p0y * (-p1x + p2x) + p0x * (p1y - p2y) + p1x * p2y);
+        },
         /**
          * Used by the containsPoint() function.
          *
@@ -408,8 +426,8 @@ var Triangle = /** @class */ (function () {
             //
             // Point-in-Triangle test found at
             //   http://stackoverflow.com/questions/2049582/how-to-determine-a-point-in-a-2d-triangle
-            //
-            var area = 1 / 2 * (-p1y * p2x + p0y * (-p1x + p2x) + p0x * (p1y - p2y) + p1x * p2y);
+            // var area : number = 1/2*(-p1y*p2x + p0y*(-p1x + p2x) + p0x*(p1y - p2y) + p1x*p2y);
+            var area = Triangle.utils.signedArea(p0x, p0y, p1x, p1y, p2x, p2y);
             var s = 1 / (2 * area) * (p0y * p2x - p0x * p2y + (p2y - p0y) * px + (p0x - p2x) * py);
             var t = 1 / (2 * area) * (p0x * p1y - p0y * p1x + (p0y - p1y) * px + (p1x - p0x) * py);
             return s > 0 && t > 0 && (1 - s - t) > 0;